refactor: Move derive_macro_expansion.rs to end of proc_macro.rs
